{{cookiecutter.project_slug}}/requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.3.1  # https://github.com/python-pillow/Pillow
{%- if cookiecutter.use_compressor == "y" %}
{%- if cookiecutter.windows == 'y' and cookiecutter.use_docker == 'n' %}
rcssmin==1.0.6 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
{%- else %}
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
{%- endif %}
{%- endif %}
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
{%- endif %}
redis==3.5.3  # https://github.com/andymccurdy/redis-py
{%- if cookiecutter.use_docker == "y" or cookiecutter.windows == "n" %}
hiredis==2.0.0  # https://github.com/redis/hiredis-py
{%- endif %}
{%- if cookiecutter.use_celery == "y" %}
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
{%- if cookiecutter.use_docker == 'y' %}
flower==0.9.7  # https://github.com/mher/flower
{%- endif %}
{%- endif %}
{%- if cookiecutter.use_async == 'y' %}
uvicorn[standard]==0.14.0  # https://github.com/encode/uvicorn
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django==3.1.13  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.45.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
{%- if cookiecutter.use_compressor == "y" %}
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-redis==5.0.0  # https://github.com/jazzband/django-redis
{%- if cookiecutter.use_drf == "y" %}
# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
{%- endif %}


{%- if cookiecutter.use_data_science_tools == "y" %}
# Data Science
# ------------------------------------------------------------------------------
scipy
numpy
pandas
matplotlib
spacy
# Machine Learning
scikit-learn
xgboost
keras
# tensorflow==1.15 # optional: lrg, may req more RAM
# joblib # optional: for pipelining/workflow
{%- endif %}

{%- if cookiecutter.use_web_crawling_tools == "y" %}
# Web crawling
# ------------------------------------------------------------------------------
bs4==0.0.1
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
selenium==3.141.0
python-anticaptcha==0.7.1
{%- endif %}

